# Make small performance refactors to callbacks