### Summary

This PR contains some refactor suggestions to ActiveSupport::Callbacks and related classes. I have split them into different commits so that if any of the suggestions is not desired, we can easily drop it. There should be no changes in functionality.

The benchmark script appears more complex than it should. It's long because I had to override a few classes. Looking at the changes per commit might make the overall intention clearer.

The benchmark was executed using Ruby 2.7.0-preview3.

Look forward to some feedback.

**Benchmark script**
```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"
  git_source(:github) { |repo| "https://github.com/#{repo}.git" }
  gem "rails", github: "rails/rails"
  gem "benchmark-ips"
end

module ActiveSupport
  module Callbacks
    def fast_run_callbacks(kind)
      callbacks = __callbacks[kind.to_sym]

      if callbacks.empty?
        yield if block_given?
      else
        env = Filters::Environment.new(self, false, nil)
        next_sequence = callbacks.compile

        # Common case: no 'around' callbacks defined
        if next_sequence.final?
          next_sequence.invoke_before(env)
          env.value = !env.halted && (!block_given? || yield)
          next_sequence.invoke_after(env)
          env.value
        else
          invoke_sequence = Proc.new do
            skipped = []

            while true
              current = next_sequence
              current.invoke_before(env)
              if current.final?
                env.value = !env.halted && (!block_given? || yield)
              elsif current.skip?(env)
                skipped << current
                next_sequence = next_sequence.nested
                next
              else
                next_sequence = next_sequence.nested
                begin
                  target, block, method, *arguments = current.expand_call_template(env, invoke_sequence)
                  target.send(method, *arguments, &block)
                ensure
                  next_sequence = current
                end
              end
              current.invoke_after(env)
              skipped.pop.invoke_after(env) while skipped.any?
              break env.value
            end
          end

          invoke_sequence.call
        end
      end
    end

    class FastCallTemplate # :nodoc:
      def initialize(target, method, arguments, block)
        @override_target = target
        @method_name = method
        @arguments = arguments
        @override_block = block
      end

      # Return the parts needed to make this call, with the given
      # input values.
      #
      # Returns an array of the form:
      #
      #   [target, block, method, *arguments]
      #
      # This array can be used as such:
      #
      #   target.send(method, *arguments, &block)
      #
      # The actual invocation is left up to the caller to minimize
      # call stack pollution.
      def expand(target, value, block)
        base_result = [@override_target || target, @override_block || block, @method_name]

        @arguments.each_with_object(base_result) do |arg, result|
          result << case arg
                    when :value; value
                    when :target; target
                    when :block; block || raise(ArgumentError)
                    end
        end
      end

      def make_lambda
        lambda do |target, value, &block|
          target, block, method, *arguments = expand(target, value, block)
          target.send(method, *arguments, &block)
        end
      end

      def inverted_lambda
        lambda do |target, value, &block|
          target, block, method, *arguments = expand(target, value, block)
          ! target.send(method, *arguments, &block)
        end
      end

      def self.build(filter, callback)
        case filter
        when Symbol
          new(nil, filter, [], nil)
        when Conditionals::Value
          new(filter, :call, [:target, :value], nil)
        when ::Proc
          if filter.arity > 1
            new(nil, :instance_exec, [:target, :block], filter)
          elsif filter.arity > 0
            new(nil, :instance_exec, [:target], filter)
          else
            new(nil, :instance_exec, [], filter)
          end
        else
          method_to_call = callback.current_scopes.join("_")

          new(filter, method_to_call, [:target], nil)
        end
      end
    end

    class FastCallback #:nodoc:#
      def self.build(chain, filter, kind, options)
        if filter.is_a?(String)
          raise ArgumentError, <<-MSG.squish
              Passing string to define a callback is not supported. See the `.set_callback`
              documentation to see supported values.
          MSG
        end

        new chain.name, filter, kind, options, chain.config
      end

      attr_accessor :kind, :name
      attr_reader :chain_config

      def initialize(name, filter, kind, options, chain_config)
        @chain_config = chain_config
        @name    = name
        @kind    = kind
        @filter  = filter
        @key     = compute_identifier filter
        @if      = check_conditionals(Array(options[:if]))
        @unless  = check_conditionals(Array(options[:unless]))
      end

      def filter; @key; end
      def raw_filter; @filter; end

      def merge_conditional_options(chain, if_option:, unless_option:)
        options = {
            if: @if.dup,
            unless: @unless.dup
        }

        options[:if].concat     Array(unless_option)
        options[:unless].concat Array(if_option)

        self.class.build chain, @filter, @kind, options
      end

      def matches?(_kind, _filter)
        @kind == _kind && filter == _filter
      end

      def duplicates?(other)
        case @filter
          when Symbol
            matches?(other.kind, other.filter)
          else
            false
          end
      end

      # Wraps code with filter
      def apply(callback_sequence)
        public_send(kind, callback_sequence)
      end

      def current_scopes
        Array(chain_config[:scope]).map { |s| public_send(s) }
      end

      def before(callback_sequence)
        user_callback = FastCallTemplate.build(@filter, self)

        Filters::Before.build(callback_sequence, user_callback.make_lambda, conditions_lambdas, chain_config, @filter)
      end

      def after(callback_sequence)
        user_callback = FastCallTemplate.build(@filter, self)

        Filters::After.build(callback_sequence, user_callback.make_lambda, conditions_lambdas, chain_config)
      end

      def around(callback_sequence)
        user_callback = FastCallTemplate.build(@filter, self)

        callback_sequence.around(user_callback, conditions_lambdas)
      end

      private

      def check_conditionals(conditionals)
        if conditionals.any? { |c| c.is_a?(String) }
          raise ArgumentError, <<-MSG.squish
                Passing string to be evaluated in :if and :unless conditional
                options is not supported. Pass a symbol for an instance method,
                or a lambda, proc or block, instead.
          MSG
        end

        conditionals
      end

      def compute_identifier(filter)
        case filter
            when ::Proc
              filter.object_id
            else
              filter
          end
      end

      def conditions_lambdas
        @if.map { |c| FastCallTemplate.build(c, self).make_lambda } +
            @unless.map { |c| FastCallTemplate.build(c, self).inverted_lambda }
      end
    end

    module ClassMethods
      def fast_set_callback(name, *filter_list, &block)
        type, filters, options = normalize_callback_params(filter_list, block)

        self_chain = get_callbacks name
        mapped = filters.map do |filter|
          FastCallback.build(self_chain, filter, type, options)
        end

        __update_callbacks(name) do |target, chain|
          options[:prepend] ? chain.prepend(*mapped) : chain.append(*mapped)
          target.set_callbacks name, chain
        end
      end
    end
  end
end

class Record
  include ActiveSupport::Callbacks
  define_callbacks :save
  define_callbacks :fast_save

  def save
    run_callbacks :save do
      true
    end
  end

  def fast_save
    fast_run_callbacks :fast_save do
      true
    end
  end
end

class OnlyBefore < Record
  set_callback :save, :before, :before_callback
  set_callback :save, :before do |obj|
    true
  end
  set_callback :save, :before, :another_before
  fast_set_callback :fast_save, :before, :before_callback
  fast_set_callback :fast_save, :before do |obj|
    true
  end
  fast_set_callback :fast_save, :before, :another_before

  def before_callback
    true
  end

  def another_before
    true
  end
end

class OnlyAfter < Record
  set_callback :save, :after, :after_callback
  set_callback :save, :after do |obj|
    true
  end
  set_callback :save, :after, :another_after
  fast_set_callback :fast_save, :after, :after_callback
  fast_set_callback :fast_save, :after do |obj|
    true
  end
  fast_set_callback :fast_save, :after, :another_after

  def after_callback
    true
  end

  def another_after
    true
  end
end

class BothTypes < Record
  set_callback :save, :before, :before_callback
  set_callback :save, :before do |obj|
    true
  end
  set_callback :save, :before, :another_before

  set_callback :save, :after, :after_callback
  set_callback :save, :after do |obj|
    true
  end
  set_callback :save, :after, :another_after

  fast_set_callback :fast_save, :before, :before_callback
  fast_set_callback :fast_save, :before do |obj|
    true
  end
  fast_set_callback :fast_save, :before, :another_before

  fast_set_callback :fast_save, :after, :after_callback
  fast_set_callback :fast_save, :after do |obj|
    true
  end
  fast_set_callback :fast_save, :after, :another_after

  def before_callback
    true
  end

  def another_before
    true
  end

  def after_callback
    true
  end

  def another_after
    true
  end
end

class AllTypes < Record
  set_callback :save, :before, :before_callback
  set_callback :save, :before do |obj|
    true
  end
  set_callback :save, :before, :another_before

  set_callback :save, :after, :after_callback
  set_callback :save, :after do |obj|
    true
  end
  set_callback :save, :after, :another_after
  set_callback :save, :around do
    true
  end

  fast_set_callback :fast_save, :before, :before_callback
  fast_set_callback :fast_save, :before do |obj|
    true
  end
  fast_set_callback :fast_save, :before, :another_before

  fast_set_callback :fast_save, :after, :after_callback
  fast_set_callback :fast_save, :after do |obj|
    true
  end
  fast_set_callback :fast_save, :after, :another_after
  fast_set_callback :fast_save, :around do
    true
  end

  def before_callback
    true
  end

  def another_before
    true
  end

  def after_callback
    true
  end

  def another_after
    true
  end
end

SCENARIOS = {
  "OnlyBefore" => OnlyBefore,
  "OnlyAfter" => OnlyAfter,
  "BothTypes" => BothTypes,
  "AllTypes" => AllTypes,
}

SCENARIOS.each_pair do |name, value|
  puts
  puts " #{name} ".center(80, "=")
  puts

  record = value.new

  Benchmark.ips do |x|
    x.report("run_callbacks")      { record.save }
    x.report("fast_run_callbacks") { record.fast_save }
    x.compare!
  end
end
```

**Output**
```bash
================================== OnlyBefore ==================================

Warming up --------------------------------------
       run_callbacks    19.001k i/100ms
  fast_run_callbacks    24.847k i/100ms
Calculating -------------------------------------
       run_callbacks    216.772k (± 8.9%) i/s -      1.083M in   5.039813s
  fast_run_callbacks    267.728k (± 3.7%) i/s -      1.342M in   5.018569s

Comparison:
  fast_run_callbacks:   267728.1 i/s
       run_callbacks:   216771.7 i/s - 1.24x  slower


================================== OnlyAfter ===================================

Warming up --------------------------------------
       run_callbacks    27.796k i/100ms
  fast_run_callbacks    37.765k i/100ms
Calculating -------------------------------------
       run_callbacks    317.756k (± 7.3%) i/s -      1.584M in   5.016030s
  fast_run_callbacks    429.472k (± 3.0%) i/s -      2.153M in   5.016629s

Comparison:
  fast_run_callbacks:   429471.9 i/s
       run_callbacks:   317755.5 i/s - 1.35x  slower


================================== BothTypes ===================================

Warming up --------------------------------------
       run_callbacks    14.191k i/100ms
  fast_run_callbacks    17.702k i/100ms
Calculating -------------------------------------
       run_callbacks    155.700k (± 5.7%) i/s -    780.505k in   5.029789s
  fast_run_callbacks    187.120k (± 3.4%) i/s -    938.206k in   5.019561s

Comparison:
  fast_run_callbacks:   187120.1 i/s
       run_callbacks:   155699.8 i/s - 1.20x  slower


=================================== AllTypes ===================================

Warming up --------------------------------------
       run_callbacks    12.784k i/100ms
  fast_run_callbacks    14.473k i/100ms
Calculating -------------------------------------
       run_callbacks    141.600k (± 5.8%) i/s -    715.904k in   5.073283s
  fast_run_callbacks    148.564k (± 3.8%) i/s -    752.596k in   5.072877s

Comparison:
  fast_run_callbacks:   148563.9 i/s
       run_callbacks:   141600.4 i/s - same-ish: difference falls within error
```